Creates packages "page", "controlextension"
