--- control.extensions/src/test/java/ControlExtensionBase.java
@@ -1,4 +1,0 @@
-
-public class ControlExtensionBase {
-
-}
