+++ control.extensions/src/test/java/ControlExtensionTests.java
@@ -1,4 +1,7 @@
 import org.testng.annotations.Test;
+
+import page.*;
+
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertNull;
