+++ control.extensions/src/test/java/controlextension/Hyperlink.java
@@ -1,0 +1,9 @@
+package controlextension;
+import org.openqa.selenium.WebElement;
+
+public class Hyperlink extends ControlExtensionBase {
+	
+	public Hyperlink(WebElement mappedElement) {
+		super(mappedElement);
+	}
+}
