+++ control.extensions/src/test/java/controlextension/RadioButton.java
@@ -1,25 +1,22 @@
+package controlextension;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
 public class RadioButton extends ControlExtensionBase {
 	
-	private WebElement mappedElement;
-	
+	private WebElement label;
 	private static String labelXPath="./../label";
 
 	public RadioButton(WebElement mappedElement) {
-		this.mappedElement = mappedElement;
+		super(mappedElement);
+		this.label = this.mappedElement.findElement(By.xpath(labelXPath));
 	}
 
 	public void select() {
-		getLabelElement().click();
+		label.click();
 	}
 	
 	public String getText() {
-		return getLabelElement().getText();
-	}
-	
-	private WebElement getLabelElement() {
-		return this.mappedElement.findElement(By.xpath(labelXPath));
+		return label.getText();
 	}
 }
