+++ control.extensions/src/test/java/controlextension/RadioGroup.java
@@ -1,22 +1,21 @@
+package controlextension;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
 public class RadioGroup extends ControlExtensionBase {
-	
-	private WebElement groupElement;
 
-	public RadioGroup(WebElement radioGroupElement) {
-		this.groupElement = radioGroupElement;
+	public RadioGroup(WebElement mappedElement) {
+		super(mappedElement);
 	}
 
 	public RadioButton getButton(String label) {
-		var buttonElement = groupElement.findElement(By.xpath("//*[text()='" +label+ "']/../input"));
+		var buttonElement = this.mappedElement.findElement(By.xpath("//*[text()='" +label+ "']/../input"));
 		
 		return new RadioButton(buttonElement);
 	}
 
 	public RadioButton getSelected() {
-		var buttonElements = groupElement.findElements(By.tagName("input"));
+		var buttonElements = this.mappedElement.findElements(By.tagName("input"));
 		
 		for(var buttonElement: buttonElements) {
 			if(buttonElement.isSelected()) {
