+++ control.extensions/src/test/java/page/LinksPage.java
@@ -1,5 +1,8 @@
+package page;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+
+import controlextension.Hyperlink;
 
 public class LinksPage {
