+++ control.extensions/src/test/java/page/PageObjectBase.java
@@ -1,3 +1,4 @@
+package page;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
