+++ control.extensions/src/test/java/page/RadioPage.java
@@ -1,6 +1,9 @@
+package page;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+
+import controlextension.RadioGroup;
 
 public class RadioPage extends PageObjectBase {
